requirements.txt
amqp==2.6.0
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.5
Django==3.0.6
django-admin-sortable2==0.7.6
django-appconf==1.0.4
django-celery-email==3.0.0
django-ckeditor==5.9.0
django-cleanup==5.0.0
django-environ==0.4.5
django-filter==2.2.0
django-js-asset==1.2.2
django-money==1.1
django-parler==2.0.1
djangorestframework==3.11.0
easy-thumbnails==2.7
future==0.18.2
gunicorn==20.0.4
importlib-metadata==1.6.0
kombu==4.6.10
Pillow==7.1.2
psycopg2-binary==2.8.5
py-moneyed==0.8.0
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
